Generated yamls with hidden attributes.

+++ project_generator/init_yaml.py
@@ -2,6 +2,7 @@ import os
 import yaml
 from .project import FILES_EXTENSIONS
 import logging
+import bisect
 
 def _determine_tool(linker_ext):
         if "sct" in linker_ext or "lin" in linker_ext:
@@ -12,7 +13,7 @@ def _determine_tool(linker_ext):
             return "iar_arm"
 
 
-def _scan(section, root, directory, extensions, is_path):
+def _scan(section, root, directory, extensions):
         if section == "sources":
             data_dict = {}
         else:
@@ -24,24 +25,27 @@ def _scan(section, root, directory, extensions, is_path):
                 if ext in extensions:
                     if section == "sources":
                         dir = directory.split(os.path.sep)[-1] if dirpath == directory else dirpath.replace(directory,'').split(os.path.sep)[1]
-                        if dir in data_dict:
-                            data_dict[dir].append(os.path.join(relpath, filename))
+                        if dir in data_dict and relpath not in data_dict[dir]:
+                            bisect.insort(data_dict[dir], relpath)
                         else:
-                            data_dict[dir] = [(os.path.join(relpath, filename))]
+                            data_dict[dir] = [(relpath)]
                     elif section == 'includes':
                         dirs = relpath.split(os.path.sep)
                         for i in range(1, len(dirs)+1):
                             data_dict.append(os.path.sep.join(dirs[:i]))
                     else:
-                        data_dict.append(relpath if is_path else os.path.join(relpath, filename))
+                        data_dict.append(os.path.join(relpath, filename))
         if section == "sources":
             return data_dict
-        return list(set(data_dict))
+        l = list(set(data_dict))
+        l.sort()
+        return l
 
 def _generate_file(filename,root,directory,data):
         logging.debug('Generating yaml file')
+        overwrite = False
         if os.path.isfile(os.path.join(directory, filename)):
-            print("Project file " + filename+ " already exists")
+            print("Project file " +filename+  " already exists")
             while True:
                 answer = raw_input('Should I overwrite it? (Y/n)')
                 try:
@@ -52,44 +56,48 @@ def _generate_file(filename,root,directory,data):
                     break
                 except ValueError:
                     continue
-        with open(os.path.join(root, filename), 'wt') as f:
-            f.write(yaml.dump(data, default_flow_style=False))
+        if overwrite:
+            with open(os.path.join(root, filename), 'r+') as f:
+                f.write(yaml.dump(data, default_flow_style=False))
+        else:
+            with open(os.path.join(root, filename), 'w+') as f:
+                f.write(yaml.dump(data, default_flow_style=False))
+        p = os.popen('attrib +h ' + filename)
+        p.close()
 
 
-def create_yaml(root, directory, project_name, board, list_sources):
+def create_yaml(root, directory, project_name, board):
         common_section = {
-            'linker_file': [False, FILES_EXTENSIONS['linker_file']],
-            'sources': [False, FILES_EXTENSIONS['source_files_c'] + FILES_EXTENSIONS['source_files_cpp'] +
-                        FILES_EXTENSIONS['source_files_s'] + FILES_EXTENSIONS['source_files_obj'] +
-                        FILES_EXTENSIONS['source_files_lib']],
-            'includes': [True, FILES_EXTENSIONS['includes']],
-            'target': [False, []],
+            'linker_file': FILES_EXTENSIONS['linker_file'],
+            'sources': FILES_EXTENSIONS['source_files_c'] + FILES_EXTENSIONS['source_files_cpp'] +
+                        FILES_EXTENSIONS['source_files_s'] + FILES_EXTENSIONS['source_files_obj'],
+            'includes': FILES_EXTENSIONS['includes'],
+            'target': [],
         }
-        projects = {
+        projects_yaml = {
             'projects': {
                 project_name: ['project.yaml']
             }
         }
 
-        data = {
+        project_yaml = {
             'common': {},
             'tool_specific': {}
         }
 
         for section in common_section:
-            if len(common_section[section][1]) > 0:
-                data['common'][section] = _scan(section, root, directory,common_section[section][1],common_section[section][0])
+            if len(common_section[section]) > 0:
+                project_yaml['common'][section] = _scan(section, root, directory,common_section[section])
 
-        data['common']['target'] = []
-        data['common']['target'].append(board)
-        tool = _determine_tool(str(data['common']['linker_file']).split('.')[-1])
-        data['tool_specific'] = {
+        project_yaml['common']['target'] = [board]
+        tool = _determine_tool(str(project_yaml['common']['linker_file']).split('.')[-1])
+        project_yaml['tool_specific'] = {
             tool: {
-                'linker_file': data['common']['linker_file']
+                'linker_file': project_yaml['common']['linker_file']
             }
         }
-        _generate_file("projects.yaml", root, directory, projects)
-        _generate_file("project.yaml", root, directory, data)
+        _generate_file("projects.yaml", root, directory, projects_yaml)
+        _generate_file("project.yaml", root, directory, project_yaml)
